+++ R/input-checkboxgroup.R
@@ -74,6 +74,8 @@
 checkboxGroupInput <- function(inputId, label, choices = NULL, selected = NULL,
   inline = FALSE, width = NULL, choiceNames = NULL, choiceValues = NULL) {
 
+  check_required(inputId, label)
+
   # keep backward compatibility with Shiny < 1.0.1 (see #1649)
   if (is.null(choices) && is.null(choiceNames) && is.null(choiceValues)) {
     choices <- character(0)