+++ R/input-textarea.R
@@ -60,6 +60,8 @@ textAreaInput <- function(
   rlang::check_dots_empty()
   updateOn <- rlang::arg_match(updateOn)
 
+  check_required(inputId, label)
+
   value <- restoreInput(id = inputId, default = value)
 
   if (!is.null(resize)) {